Reorg Util Module to .db and .net modules
